Last round of checkin introduced a bug.  This checkin corrected it and made a few minor changes.
Last round of check in fixed the problem when an attribute was opened twice and data was 
written with one of the handles, the file didn't have the data.  It happened because each 
handle had its own object structure, and the empty one overwrote the data with fill value.
This is fixed by making some attribute information like the data be shared in the
attribute structure.

Tested on smirom, kagiso, and linew.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15390 dab4d1a6-ed17-0410-a064-d1ae371a2980