Purpose: Update Windows test script for new h5diff option: -c
Description:
A new parameter was added to h5diff recently.  We add a new test to the h5diff script, and change the syntax in which h5diff is called from the h5repack test script.  The output for one of the h5repack tests also needed to be changed.

Tested:
VS6 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15449 dab4d1a6-ed17-0410-a064-d1ae371a2980