Purpose: Add Windows Fortran DLL export code to separate DEF file.
Description:
In in Fortran source code, there was a great deal of code that was necessary for Windows DLLs, and ignored for others systems.  To remove some of the bloat in the source code, we moved these definitions into separate *.def file, which will be used on by the Windows DLL project.

Tested:
VS2005 on WinXP

Note: The Windows project file will still need to be edited-- I will check that in soon.

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15583 dab4d1a6-ed17-0410-a064-d1ae371a2980