Purpose: Add h5repack alignment test to Windows, and disable ublock test
Description:
revision 15679 added a test to h5repack for alignment.  We've now added the test to the Windows script as well.

This checkin also disables a previously-added h5repack test for the userblock.  The test is disabled for now because Windows cannot find the input file ublock.bin.  The file is generated dynamically on other systems, but cannot be easily generated for use on Windows.  We should re-enable this test in the future when ublock.bin is added to svn.

Tested:
VS6 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15717 dab4d1a6-ed17-0410-a064-d1ae371a2980