Purpose: Update Windows h5repack test script for new directory
Description:
The h5repack test files were added to a new testfiles/ subdirectory, and the test script update.  This checkin brings those changes to Windows.  We also re-enable a test for the userblock which was disabled for other reasons before.

Note: One h5repack test is still failing, but I think it's a minor error that Pedro and I can correct fairly easily later today.

Tested
VS2005 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15783 dab4d1a6-ed17-0410-a064-d1ae371a2980