Purpose: Bring recent Windows-specific fixes from 1.8 branch to trunk
Description:
This checkin includes revisions 15845, 15846, and 15853 from the 1.8 branch.  From the svn log:

r15845:
Purpose: Add Windows-specific version of HDftruncate

Description:
Windows doesn't include a version of the system call ftruncate.  There is a similar call, _chsize_s, which performs very similarly.  Thus, we map HDftruncate to _chsize_s in our Windows header file, H5win32defs.h.

r15846:
Purpose: Declare loop variable at beginning of function

Description:
On Windows, we were getting compile errors because h5test.c includes a function with variable declarations in the middle of the call.  The Microsoft C compiler demands that all variables be declared at the top of the function.  This checkin simply moves the declaration to the top to satisfy Visual Studio.

r15853:
Purpose: Decrease size of points array in links test

Description:
In the external_set_elink_fapl2 links test, there is a large array of points which declared on the stack for testing.  Previously, the array was 1000 x 1000, which was too large for Visual Studio to handle.  As a result, we were getting stack overflows during the test.  We've reduced the number to 400, as this seems to be below the limit.  The exact number of points in the array isn't important to the test.

Tested:
VS2005 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15855 dab4d1a6-ed17-0410-a064-d1ae371a2980