Description:
We check for all the available reals in Fortran and if 16 byte real is available in Fortran and not in C then we disable the 16 byte real in Fortran. Also added the test for 12 byte real in Fortran so that it can match the 12 byte float in C if available. Note: if KIND=10 and KIND=16 are both avaiable as when using g95, then it may be the case on some systems that the size of KIND=10 and KIND=16 are both 16 bytes, so the program will print twice in H5fort_type_defines.h 

#define H5_FORTRAN_HAS_REAL_16

which should not cause any errors.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@15864 dab4d1a6-ed17-0410-a064-d1ae371a2980