Purpose: Fix Windows definition of H5ftruncate for VS.NET
Description:
On Windows/Visual Studio, there is no function ftruncate.  As a result, we have redefined the H5ftruncate macro to use _chsize_s.  However, this function also doesn't exist on Visual Studio prior to 2005.  Previously we had an #ifdef to support VS 6.0 with the lesser chsize function.  This checkin changes the #ifdef to also include VS.NET, since we still support it for HDF5 1.8 branch.

Tested:
VS.NET on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@16019 dab4d1a6-ed17-0410-a064-d1ae371a2980