Purpose: Add functions to allow more flexible traversal of external links.
Description: Added H5Pset/get_elink_cb to allow the user to specify a callback
function to be called whenever an external link is traversed.  Added
H5Pset/get_elink_acc_flags to allow the user to specify the file access flags
to use to open the target file of an external link.  All these properties are set on a LAPL.

Tested: jam (trunk version tested with h5committest)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@16282 dab4d1a6-ed17-0410-a064-d1ae371a2980