Description: 	Bring r17402 & r17403 from trunk to 1.8 branch:
r17402:
    Switch from using a 'flags' parameter to the flush routine(s), to just
using a single 'closing' boolean parameter, since that's all we're doing with
the flags anyway and this makes things more obvious.

r17403:
    Eliminate 'closing' flag to dataset flush routine, since all cached
dataset information has already been flushed when a file is closed and the
datasets themselves will be closed (and the dataset caches destroyed).

    Skip calling more flush routines when the file is closing, in preparation
for eventual separation of the 'flush' concept from the 'shutdown the cache'
concept.

Tested on:
	FreeBSD/32 6.3 (duty) in debug mode
	(h5committest performed on trunk)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@17404 dab4d1a6-ed17-0410-a064-d1ae371a2980