Bug fix and improvement.
Description:
Build failed when embedded-info is not enabled. Fixed.
testlibinfo.sh output was ugly, did not prepare the case of embedded-info
not enabled. Fixed.  Also changed NOT to test test binary even if static-exec
is enabled.

Tested:
jam (with and without embedded-info enabled).  Did not test in other
platforms since these were script changes that are platform independent.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@17500 dab4d1a6-ed17-0410-a064-d1ae371a2980