Added parallel tests for rank projection code, found and fixed bug in H5Smpio.c, added test code to expose bug in H5Smpio.c that doesn't depend on rank projection code.
Also encountered what may be a bug in MPICH -- expresses itself as a
crash in a call to MPI_Type_commit() just after the final call to 
MPI_Type_struct() in the construction of the derived type intended to
handle initialization of a test chunked data set (in 
H5D_link_chunk_collective_io()).  Some suggestion that empty sub-derived
types (for chunks in which no I/O is required for this process) may be
an issue.  I was able to cause the bug to appear and disappear by changing
chunk dimensions.  

I have not tracked down the exact cause of this failure.

Thus this revision DOES NOT run on all tested platforms.  Specifically
it passes on some systems -- i.e.:

        MPICH 1.2.7 (old version from 2005) on 64 bit Linux / Opteron 
	(phoenix)

        MPICH 1.2.7 (old version from 2005) on 32 bit Linux / Athlon-MP
	(mars)

        MPICH 1.1 (also an old version) on 64 bit Free BSD / Opteron
        (some mpd hangs here -- but runs eventually) (liberty)

and fails on others:

        MPICH 1.0.8 on 32-bit Linux / Xeon
        (jam)

        MPICH 1.2 on 32-bit Linux / Xeon
        (jam)

        MPICH 1.2.1 on 32-bit Linux / Xeon
        (jam)

Note that all these tests were parallel with debugging enabled.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@17947 dab4d1a6-ed17-0410-a064-d1ae371a2980