Checking in fix for MPI type handling bug reported by Rob Latham.  Had hoped to modify the existing test for the H5S_obtain_datatype() bug to test the fix to Rob's bug as well, but was unable to do this cleanly. Thus I will have to make another checkin with Rob's test case adapted to a automatic regression test.
Also added some barriers between the writes and reads of initial values of
test data sets in the rank projection and H5S_obtain_datatype() tests.
Added them in the hope that they would address the failures on tests of 
initial value of data sets with different versions of MPI that both 
Rob and I have observed, but no luck.  Left them in as the don't hurt,
and I'm not completely convinced they are not needed.

Tested (and passed) on Phoenix (Opteron/Linux) and Mars (Athlon-MP/Linux).
Tested (and failed) on Jam with MPICH 1.2.1 (this due to known bug in 
MPICH).

All tests parallel with debugging enabled.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@18017 dab4d1a6-ed17-0410-a064-d1ae371a2980