Purpose: Improve performance of the the chunk cache
Description:
Previously, the chunk cache would, when looking for a chunk, alwaylook for the
chunk in the b-tree before checking the cache.  Reworked the functions that
look up chunks to always check the cache first and avoid the b-tree lookup if
the chunk is in cache.

Tested: jam, linew, amani (h5committest)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19006 dab4d1a6-ed17-0410-a064-d1ae371a2980