bug 1678: temporary fix.
Description: aix -q32 default to a memory limit that some HDF5 tests would
fail in malloc.

Solution:
Temporary hardset
RUNSERIAL=${RUNSERIAL="env LDR_CNTRL=MAXDATA=0x20000000@DSA"}
to bypass this problem until the test programs are fixed.  Though -q64 binary
has a higher default memory limit and does not need this patch, it is applied
the same for the sake of simplicity since it is a temporary fix.

Tested:
in uP, tested 32 and -q64 modes. No h5committest since this applies to the
AIX platform only.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19303 dab4d1a6-ed17-0410-a064-d1ae371a2980