In this revision the '/' after ${DESTDIR} wherever it occurs has been removed. According to the official description of DESTDIR by Gnu at http://www.gnu.org/prep/standards/html_node/DESTDIR.html, DESTDIR is prepended to the normal and complete install path that it precedes for the purpose of installing in a temporary directory which is useful for building rpms and other packages.  The '/' after ${DESTDIR} will be followed by another '/' at the beginning of the normal install path.  When DESTDIR is empty the path then begins with '//', which is incorrect and causes problems at least for Cygwin.
The change made in config/commence.am affects many Makefile.*s when bin/reconfigure
is run.  EXAMPLEDIRs had to be corrected individually.

Also the scripts to run examples after they are installed have checks added to see 
that the fortran and c++ directories are present before trying to run the scripts 
in them.

Tested with h5committest on amani, heiwa and jam.




git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19408 dab4d1a6-ed17-0410-a064-d1ae371a2980