Bug fix: 1961-- AIX 6.1 --enable-share did not work.
Description:
The AIX platform does not install the lib*.so for some reason but libhdf5.a
acts like a shared library.  Until we can find out how to fix this, patch
up tetlibinfo.sh to not look for an lib*.so in AIX. (Note the test results
look wierd as it says libhdf5.a passed and skipped.)

Tested: NCSA BP for both enable and disable-shared.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19462 dab4d1a6-ed17-0410-a064-d1ae371a2980