Purpose: Fix bug 1864
Description:
Library versions 1.6.3 and earlier contain a bug which causes them to be unable
to perform certain operations on a group if that group's symbol table
information is not cached in the parent group's symbol table.  Versions 1.8.0
to 1.8.5 did not cache this information.  Modified library to cache this
information.

Tested: jam, amani, heiwa (h5committest)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19464 dab4d1a6-ed17-0410-a064-d1ae371a2980