Modfied object header cache client code to update the contents of the in core image of the on disk representation of the object header on clear as well as on flush.  Made this change only in the parallel case, as it causes problems only there, and only with the round robin metadata write strategy in which an object header may be flushed by several different processes in the course of a computation, and marked clean on all other processes on each flush.
Note that I made this change to both the H5O_clear() and 
H5O_cache_chk_clear() cache client callbacks.  At present I don't think 
my fix in H5O_cache_chk_clear() does anything, as I don't think it is 
possible to dirty a chunk without also marking the associate header dirty 
as well.  However I will leave the fix in as the client code is presently 
in a state of flux, and it may be a useful reminder.

Added test to t_mdset.c to verify the fix.

Also added test code to H5C.c and H5AC.c which dumps a listing of all
entries in the cache, along with their vital statistics.  This code was
needed to rule out the possibility of the bug being caused by a failure
in the metadata cache -- I leave it in as we may want it again some time. 



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@19616 dab4d1a6-ed17-0410-a064-d1ae371a2980